authentication_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
jwt
python-dotenv
pika
bcrypt==4.0.1
python-jose[cryptography]
redis